[#47] refactor: 간단한 코드 정리
